// types/extended.ts
export interface ProjectInfo extends Project{
  id: string;
  title: string;
}

export type TaskType =   {
  id: string
  position : string | null,
  workspaceId: string
  creator: UserInfo;
  name: string,
  taskStatus: TaskStatus,
  project: ProjectInfo,
  title :         string
  emoji:          string  
  updatedUserId:          string  
  dueDate?: string,
}
// ...
export interface UserInfo {
  id: string;
  username: string;
  image?: string | null;
  name?: string | null;
  surname?: string | null;
}